# api/openapi.yaml
openapi: 3.0.3
info:
  title: LatentTone Product API
  description: |
    Phase 2+3 HTTP product surface for LatentTone (auth, sessions, stream/HLS,
    feedback, neighbor playlists, Phase 3C user playlists, embed status).

    **Try it out (Swagger / scripts):** call `POST /api/v1/auth/login` (or register),
    copy the response `token`, click **Authorize**, and paste it as a Bearer token.
    The product SPA uses the HttpOnly cookie `lt_session` only (same-origin; no
    Bearer in JavaScript). Bearer remains supported for non-browser clients.

    SPA and API share one origin (`/app/` + `/api/v1/`) — no CORS. Do not add
    `Access-Control-Allow-Origin: *`.

    Swagger UI is flag-gated (`enable_api_docs`) and is **not** a substitute for
    Phase 3 stream-smoke Gates A–D or `/dev/stream`.
  version: 0.3.1
  contact:
    name: martinsah
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: /
    description: Same origin as latenttone serve
  - url: https://latent.lt.lkeng.org
    description: Example reverse-proxied public origin (override via PUBLIC_BASE_URL)
...
paths:
...
  /api/v1/sessions:
    post:
      tags: [sessions]
      summary: Create listening session
      description: |
        Starts a session from `seed_track_id`. Returns short-poll status including
        `now_playing`, `queue`, `hls_url`, and `progressive_url`.

        Warning: long-lived sessions hold HLS disk under `hls_root`; always DELETE
        when finished. Swagger Try-it is for contract probing, not Gate B listen verification.
      operationId: createSession
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              seed_track_id: 1
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '503':
          description: Too many concurrent sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/sessions/{id}:
    get:
      tags: [sessions]
      summary: Session status (short-poll)
      description: Poll for `now_playing`, `queue`, and `last_feedback`.
      operationId: getSession
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      tags: [sessions]
      summary: Stop session
      description: Stops the session and cleans up session HLS files when possible.
      operationId: stopSession
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        '200':
          description: Stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/sessions/{id}/feedback:
    post:
      tags: [feedback]
      summary: Apply feedback signal
      description: |
        Signals: `like`, `dislike`, `skip`, `ban`, `complete` (natural end).
        Optional `track_id` defaults to
        the session's current now-playing track. Response is the updated session status
        (field names match `stream_smoke.sh`).
      operationId: postFeedback
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
            example:
              signal: skip
              track_id: 1
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/sessions/{id}/queue:
    post:
      tags: [sessions]
      summary: Inject track into upcoming queue (play next)
      description: |
        Phase 3 V5b / ADR-007. Pins `track_id` at `position` `next` (default) or `end`.
        Duplicates return 409. Response is the updated session status (`queue[].source`
        is `user_pin` for pinned entries).
      operationId: postQueueInject
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [track_id]
              properties:
                track_id:
                  type: integer
                  format: int64
                position:
                  type: string
                  enum: [next, end]
                  default: next
            example:
              track_id: 418
              position: next
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /api/v1/sessions/{id}/back:
    post:
      tags: [sessions]
      summary: Go to previous track in session history
      description: |
        Restores the previous played/skipped track and pushes the current track
        to the front of the queue. Returns 409 when history is empty (client should
        restart the current track from 0). Does not stop the station.
      operationId: postSessionBack
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStatus'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
  /api/v1/me/stream-prefs:
    get:
      tags: [stream]
      summary: Get stream encode preferences
      description: |
        Per-user progressive/HLS defaults. Missing rows return `stream_format=original`
        and `bitrate_kbps=192`. `original` serves the file when the container is
        browser-safe; otherwise the server auto-transcodes to MP3. `mp3` / `aac`
        always transcode at the configured bitrate.
      operationId: getStreamPrefs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamPrefs'
        '401':
          $ref: '#/components/responses/Error'
    patch:
      tags: [stream]
      summary: Update stream encode preferences
      description: Partial update; omitted fields keep their current (or default) values.
      operationId: patchStreamPrefs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamPrefsPatch'
            example:
              stream_format: mp3
              bitrate_kbps: 192
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamPrefs'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque session id from login/register JSON field `token`.
        Preferred for Swagger Try it out: Authorize → Bearer `<token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: lt_session
      description: |
        HTTP-only cookie set on register/login. Same-origin browser clients use
        this automatically; Swagger Try-it defaults to Bearer instead.
  parameters:
    SessionID:
      name: id
      in: path
      required: true
      description: Listening session id (opaque)
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    StreamPrefs:
      type: object
      description: |
        Progressive/HLS encode defaults. Default `stream_format=original`,
        `bitrate_kbps=192`.
      properties:
        user_id:
          type: integer
          format: int64
        stream_format:
          type: string
          enum: [original, mp3, aac]
        bitrate_kbps:
          type: integer
          minimum: 64
          maximum: 320
        updated_at:
          type: string
          format: date-time
      required: [user_id, stream_format, bitrate_kbps]
    StreamPrefsPatch:
      type: object
      properties:
        stream_format:
          type: string
          enum: [original, mp3, aac]
        bitrate_kbps:
          type: integer
          minimum: 64
          maximum: 320
...
    CreateSessionRequest:
      type: object
      required: [seed_track_id]
      properties:
        seed_track_id:
          type: integer
          format: int64
          minimum: 1
...
    TrackRef:
      type: object
      properties:
        track_id:
          type: integer
          format: int64
        score:
          type: number
          format: double
        source:
          type: string
          description: Set to user_pin when manually injected (V5b)
          enum: [user_pin]
        feedback:
          type: string
          description: Latest like/dislike signal for the session user (when present)
          enum: [like, dislike]
        play_count:
          type: integer
          format: int64
          description: Global playback_events count for the track
    FeedbackAck:
      type: object
      properties:
        signal:
          type: string
          enum: [like, dislike, skip, ban, complete]
        track_id:
          type: integer
          format: int64
        at:
          type: string
          description: Server timestamp string
    SessionStatus:
      type: object
      description: Short-poll status DTO (aligned with session.StatusDTO / stream_smoke.sh)
      properties:
        id:
          type: string
        user_id:
          type: integer
          format: int64
        status:
          type: string
        seed_track_id:
          type: integer
          format: int64
        now_playing:
          $ref: '#/components/schemas/TrackRef'
        history:
          type: array
          description: Recently played/skipped tracks (oldest → newest), capped (e.g. 8)
          items:
            $ref: '#/components/schemas/TrackRef'
        queue:
          type: array
          items:
            $ref: '#/components/schemas/TrackRef'
        last_feedback:
          $ref: '#/components/schemas/FeedbackAck'
        hls_url:
          type: string
          example: /api/v1/sessions/{id}/hls/index.m3u8
        progressive_url:
          type: string
          example: /api/v1/tracks/{id}/stream
        can_go_back:
          type: boolean
          description: True when session history has at least one previous track
    FeedbackRequest:
      type: object
      required: [signal]
      properties:
        signal:
          type: string
          enum: [like, dislike, skip, ban, complete]
          description: |
            `complete` = natural track end (client `ended`); advances without
            session-skip / affinity penalty.
        track_id:
          type: integer
          format: int64
          description: Defaults to now-playing when omitted or zero
...
security:
  - bearerAuth: []